third_party/blink/web_tests/fast/xpath/invalid-functions-expected.txt

PASS document.createExpression("foobar()", null) threw exception SyntaxError: Failed to execute 'createExpression' on 'Document': The string 'foobar()' is not a valid XPath expression..
PASS document.evaluate("foobar()", document, null, XPathResult.ANY_TYPE, null) threw exception SyntaxError: Failed to execute 'evaluate' on 'Document': The string 'foobar()' is not a valid XPath expression..
PASS document.createExpression("position(//div)", null) threw exception SyntaxError: Failed to execute 'createExpression' on 'Document': The string 'position(//div)' is not a valid XPath expression..
PASS document.evaluate("position(//div)", document, null, XPathResult.ANY_TYPE, null) threw exception SyntaxError: Failed to execute 'evaluate' on 'Document': The string 'position(//div)' is not a valid XPath expression..
PASS document.createExpression("ceiling(1,2)", null) threw exception SyntaxError: Failed to execute 'createExpression' on 'Document': The string 'ceiling(1,2)' is not a valid XPath expression..
PASS document.evaluate("ceiling(1,2)", document, null, XPathResult.ANY_TYPE, null) threw exception SyntaxError: Failed to execute 'evaluate' on 'Document': The string 'ceiling(1,2)' is not a valid XPath expression..
PASS document.createExpression("ceiling(1,2)", null).evaluate(document, XPathResult.ANY_TYPE, null) threw exception SyntaxError: Failed to execute 'createExpression' on 'Document': The string 'ceiling(1,2)' is not a valid XPath expression..
PASS document.createExpression("boolean()", null) threw exception SyntaxError: Failed to execute 'createExpression' on 'Document': The string 'boolean()' is not a valid XPath expression..
PASS document.evaluate("boolean()", document, null, XPathResult.ANY_TYPE, null) threw exception SyntaxError: Failed to execute 'evaluate' on 'Document': The string 'boolean()' is not a valid XPath expression..
PASS successfullyParsed is true

TEST COMPLETE
